% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils.R
\docType{data}
\name{pH_data}
\alias{pH_data}
\title{Script to process soils data across sites
Sites: Peat 20 yr old, Alluvium 4 yr old. More to follow}
\format{An object of class \code{data.frame} with 60 rows and 9 columns.}
\usage{
pH_data
}
\description{
Script to process soils data across sites
Sites: Peat 20 yr old, Alluvium 4 yr old. More to follow
}
\keyword{datasets}
